requirements.txt
deepgram-sdk
pyaudio
sqlalchemy
psycopg2-binary
python-dotenv
